Former-commit-id: c6da746a213a4e9c58ddbbaf6bb7a4a969bb2e62
Former-commit-id: 4730fb1c0108e93dce5ca753a3bab0c8d4a6befa
Former-commit-id: 83a393b73827b0b2e222770f89fb2a8b4024babc

+++ website/php/WebsiteAPI.php
@@ -1,6 +1,10 @@
 <?php
 
+date_default_timezone_set('America/New_York');
+
 require_once 'API.class.php';
+require_once '../lib/swiftmailer/lib/swift_required.php';
+
 class WebsiteAPI extends API
 {
 
@@ -8,6 +12,8 @@ class WebsiteAPI extends API
 	private $mysqli = NULL;
 
 	public function __construct($request, $origin) {
+		$this->config = parse_ini_file("../../halite.ini", true);
+
 		$this->initDB();
 
 		$this->sanitizeHTTPParameters();
@@ -24,13 +30,16 @@ class WebsiteAPI extends API
 		}
 	}
 
+	private function encryptPassword($password) {
+		return $this->mysqli->real_escape_string(crypt($password, $this->config['encrypt']['salt']));
+	}
+
 	// Initializes and returns a mysqli object that represents our mysql database
 	private function initDB() {
-		$config = parse_ini_file("../../halite.ini");
-		$this->mysqli = new mysqli($config['hostname'],
-			$config['username'],
-			$config['password'],
-			$config['databaseName']);
+		$this->mysqli = new mysqli($this->config['database']['hostname'],
+			$this->config['database']['username'],
+			$this->config['database']['password'],
+			$this->config['database']['name']);
 
 		if (mysqli_connect_errno()) {
 			echo "<br><br>There seems to be a problem with our database. Reload the page or try again later.";
@@ -63,7 +72,8 @@ class WebsiteAPI extends API
 	protected function user() {
 		if(isset($_GET["username"])) {
 			if(isset($_GET["password"])) {
-				return $this->select("SELECT * FROM User WHERE username = '{$_GET['username']}' AND password = '{$_GET['password']}'");
+				$password = $this->encryptPassword($_GET['password']);
+				return $this->select("SELECT * FROM User WHERE username = '{$_GET['username']}' AND password = '$password'");
 			} else {
 				$fields = $this->select("SELECT * FROM User WHERE username = '{$_GET['username']}'");
 				unset($fields["password"]);
@@ -71,7 +81,8 @@ class WebsiteAPI extends API
 			}
 		} else if (isset($_GET["userID"])) {
 			if(isset($_GET["password"])) {
-				return $this->select("SELECT * FROM User WHERE userID = '{$_GET['userID']}' AND password = '{$_GET['password']}'");
+				$password = $this->encryptPassword($_GET['password']);
+				return $this->select("SELECT * FROM User WHERE userID = '{$_GET['userID']}' AND password = '{$password}'");
 			} else {
 				$fields = $this->select("SELECT * FROM User WHERE userID = '{$_GET['userID']}'");
 				unset($fields["password"]);
@@ -81,22 +92,48 @@ class WebsiteAPI extends API
 			$results = $this->selectMultiple("SELECT * FROM User WHERE status = 3");
 			foreach(array_keys($results) as $key) unset($results[$key]["password"]);
 			return $results;
+		} else if(isset($_POST['verificationCode']) && isset($_POST['userID'])) {
+			$user = $this->select("SELECT verificationCode FROM User WHERE userID={$_POST['userID']} LIMIT 1");
+			if($user['verificationCode'] == $_POST['verificationCode']) {
+				$this->insert("UPDATE User SET isVerified=1 WHERE userID={$_POST['userID']}");
+				return "Success";
+			}
+			return "Fail";
 		} else if (isset($_POST["username"]) && isset($_POST["email"]) && isset($_POST["password"])) {
 			$username = $_POST["username"];
 			$email = $_POST["email"];
-			$password = $_POST["password"];
+			$password = $this->encryptPassword($_POST["password"]);
 
 			$usernameArray = $this->select("SELECT username FROM User WHERE username = '$username' LIMIT 1");
 			if(isset($usernameArray['username'])) {
 				return "Username already exists";
 			}
-			
+
 			$emailArray = $this->select("SELECT email FROM User WHERE email = '$email' LIMIT 1");
 			if(isset($emailArray['email'])) {
 				return "Email already exists";
 			}
 
-			$this->insert("INSERT INTO User (username, email, password, mu, sigma, status) VALUES ('$username', '$email', '$password', 25.000, 8.333, 0)");
+			// Send verification email
+			$verificationCode = rand(0, 9999999999);
+			try{
+				$transport = Swift_SmtpTransport::newInstance("smtp.gmail.com", 465, "ssl")
+				->setUsername($this->config['email']['email'])
+				->setPassword($this->config['email']['password']);
+				$mailer = Swift_Mailer::newInstance($transport);
+				$message = Swift_Message::newInstance("Halite Email Verification")
+					->setFrom(array($this->config['email']['email'] => "Halite Competition"))
+					->setTo(array($email));
+
+				$this->insert("INSERT INTO User (username, email, password, mu, sigma, status, verificationCode) VALUES ('$username', '$email', '$password', 25.000, 8.333, 0, '$verificationCode')");
+				$userID = $this->select("SELECT userID FROM User WHERE email='$email' LIMIT 1")['userID'];
+
+				$message->setBody("To verify you email, <a href='halite.io/website/index.php?verificationCode={$verificationCode}&userID={$userID}'>click here</a>.", 'text/html');
+				$result = $mailer->send($message);
+			} catch (Exception $e) {
+				return "Invalid email address";
+			}
+
 			return "Success";
 		}
 	}
@@ -134,6 +171,11 @@ class WebsiteAPI extends API
 		if(isset($_FILES['botFile']['name']) && isset($_POST['userID'])) {
 			$userID = $_POST['userID'];
 
+			$user = $this->select("SELECT isVerified FROM User WHERE userID={$_POST['userID']}");
+			if(count($user) == 0 || $user['isVerified'] == false) {
+				return "Unverified email";
+			}
+
 			$targetPath = "../../storage/bots/{$userID}.zip";
 			if(file_exists($targetPath) == true) unlink($targetPath);
 
@@ -152,14 +194,23 @@ class WebsiteAPI extends API
 			else return NULL;
 		} else if(isset($_POST['username']) & isset($_POST['password'])) {
 			$username = $_POST['username'];
-			$password = $_POST['password'];
-			$_SESSION = $this->select("SELECT * FROM User WHERE username = '$username' AND password = '$password'");
+			$password = $this->encryptPassword($_POST['password']);
+
+			$user = $this->select("SELECT * FROM User WHERE username = '$username' AND password = '$password'");
+			if($user['isVerified'] == false) {
+				return "Unverified user";
+			}
+			$_SESSION = $user;
 			return "Success";
 		} else if(isset($_POST['userID']) & isset($_POST['password'])) {
 			$userID = $_POST['userID'];
-			$password = $_POST['password'];
+			$password = $this->encryptPassword($_POST['password']);
 
-			$_SESSION = $this->select("SELECT * FROM User WHERE userID = $userID AND password = '$password'");
+			$user = $this->select("SELECT * FROM User WHERE userID = $userID AND password = '$password'");
+			if($user['isVerified'] == false) {
+				return "Unverified user";
+			}
+			$_SESSION = $user;
 			return "Success";
 		} else if($this->method == 'DELETE') {
 			session_destroy();
